src/requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1607109633063/work
ansi2html==1.6.0
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828452693/work
arrow==0.15.5
astroid==2.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
binaryornot==0.4.4
bleach @ file:///tmp/build/80754af9/bleach_1611690509631/work
blinker==1.4
blis==0.7.4
boto3==1.16.61
botocore==1.19.61
Brotli @ file:///tmp/build/80754af9/brotli-split_1602517619121/work
brotlipy==0.7.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
catalogue==1.0.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255099073/work
chardet==4.0.0
chevron==0.13.1
click==7.1.1
colorama==0.4.3
cookiecutter==1.6.0
cryptography==2.9.2
cvxopt==1.2.4
cycler==0.10.0
cymem==2.0.5
dash==1.12.0
dash-bootstrap-components==0.10.2
dash-core-components==1.10.0
dash-cytoscape==0.2.0
dash-daq==0.5.0
dash-editor-components==0.0.2
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
dateparser==0.7.4
decorator==4.4.2
defusedxml==0.6.0
dill @ file:///tmp/build/80754af9/dill_1605066453875/work
docker==4.2.0
entrypoints==0.3
explainx==2.406
Flask==1.1.2
Flask-Compress @ file:///tmp/build/80754af9/flask-compress_1605067594328/work
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gcloud==0.17.0
gevent @ file:///tmp/build/80754af9/gevent_1609093112907/work
gmpy2==2.0.8
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
googleapis-common-protos @ file:///tmp/build/80754af9/googleapis-common-protos-feedstock_1600455614392/work
greenlet==1.0.0
grpcio==1.32.0
h5py @ file:///tmp/build/80754af9/h5py_1593454119955/work
heatmapz==0.0.4
httplib2==0.18.1
idna==2.8
imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1611258163564/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata==1.6.0
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
intel-openmp==2021.1.1
interpret==0.2.2
interpret-core==0.2.2
ipykernel==5.4.3
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
isort==4.3.21
itsdangerous==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
Jinja2==2.11.1
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.0.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148959479/work
jupyter-dash==0.2.1.post1
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jws==0.1.3
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy @ file:///tmp/build/80754af9/lazy-object-proxy_1607707315973/work
lime==0.2.0.1
llvmlite==0.35.0
mahotas==1.4.11
Markdown @ file:///tmp/build/80754af9/markdown_1605111062842/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
matplotlib==3.3.4
mccabe==0.6.1
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl==2021.1.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools @ file:///tmp/build/80754af9/more-itertools_1605111547926/work
mpldatacursor==0.7.1
mpmath==1.1.0
multidict @ file:///tmp/build/80754af9/multidict_1600456400975/work
murmurhash @ file:///tmp/build/80754af9/murmurhash_1607456104392/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914821128/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook @ file:///tmp/build/80754af9/notebook_1607976307905/work
numba==0.52.0
numexpr==2.7.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauth2client==3.0.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pandas==1.2.1
pandasql==0.7.3
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parso==0.8.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
plac==1.1.3
plotly==4.5.1
pluggy==0.13.1
poyo==0.5.0
preshed==3.0.5
progressbar2==3.53.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit==3.0.14
protobuf==3.14.0
psutil==5.8.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pvlib==0.7.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.4.3
pydub==0.24.1
pygam==0.8.0
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyJWT @ file:///tmp/build/80754af9/pyjwt_1610911319507/work
pylint==2.5.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PyPDF2==1.26.0
Pyrebase==3.0.27
pyrsistent==0.16.0
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
pytest==6.2.2
pytest-mock==3.0.0
python-dateutil==2.8.0
python-jwt==2.0.1
python-utils==2.5.5
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
resampy==0.2.2
retrying==1.3.3
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
s3transfer==0.3.4
SALib==1.3.12
scaleogram==0.9.5
scikit-image==0.18.1
scikit-learn==0.24.1
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
serverlessrepo==0.1.9
shap==0.37.0
six @ file:///tmp/build/80754af9/six_1605205313296/work
slicer==0.0.3
spacy==2.3.5
SQLAlchemy==1.3.22
srsly @ file:///tmp/build/80754af9/srsly_1607548514031/work
sympy @ file:///tmp/build/80754af9/sympy_1608137658835/work
tables==3.6.1
tbb==2021.1.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-addons==0.12.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
thinc @ file:///tmp/build/80754af9/thinc_1607710140838/work
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.2
tomlkit==0.5.8
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm @ file:///tmp/build/80754af9/tqdm_1593446365756/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
treeinterpreter==0.2.3
typed-ast==1.4.1
typeguard==2.10.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
tzlocal==2.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
wasabi @ file:///tmp/build/80754af9/wasabi_1599244993840/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client @ file:///tmp/build/80754af9/websocket-client_1602006590740/work
Werkzeug==1.0.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1598045053219/work
zipp==3.1.0
zope.event==4.5.0
zope.interface @ file:///tmp/build/80754af9/zope.interface_1606940237377/work
